requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
drf-yasg>=1.21
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
